# app/controllers.go
// Code generated by goagen v1.2.0-dirty, DO NOT EDIT.
//
// API "todoz": Application Controllers
//
// Command:
// $ goagen
// --design=github.com/konchanSS/Todoz/design
// --out=$(GOPATH)/src/github.com/konchanSS/Todoz
// --version=v1.2.0-dirty

package app

import (
	"context"
	"github.com/goadesign/goa"
	"github.com/goadesign/goa/cors"
	"net/http"
)

// initService sets up the service encoders, decoders and mux.
func initService(service *goa.Service) {
	// Setup encoders and decoders
	service.Encoder.Register(goa.NewJSONEncoder, "application/json")
	service.Encoder.Register(goa.NewGobEncoder, "application/gob", "application/x-gob")
	service.Encoder.Register(goa.NewXMLEncoder, "application/xml")
	service.Decoder.Register(goa.NewJSONDecoder, "application/json")
	service.Decoder.Register(goa.NewGobDecoder, "application/gob", "application/x-gob")
	service.Decoder.Register(goa.NewXMLDecoder, "application/xml")

	// Setup default encoder and decoder
	service.Encoder.Register(goa.NewJSONEncoder, "*/*")
	service.Decoder.Register(goa.NewJSONDecoder, "*/*")
}

// TodosController is the controller interface for the Todos actions.
type TodosController interface {
	goa.Muxer
	Create(*CreateTodosContext) error
	Delete(*DeleteTodosContext) error
	DeleteAll(*DeleteAllTodosContext) error
	List(*ListTodosContext) error
	Show(*ShowTodosContext) error
	Update(*UpdateTodosContext) error
}

// MountTodosController "mounts" a Todos resource controller on the given service.
func MountTodosController(service *goa.Service, ctrl TodosController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/todos/", ctrl.MuxHandler("preflight", handleTodosOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/todos/:id", ctrl.MuxHandler("preflight", handleTodosOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCreateTodosContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Create(rctx)
	}
	h = handleTodosOrigin(h)
	service.Mux.Handle("POST", "/todos/", ctrl.MuxHandler("create", h, nil))
	service.LogInfo("mount", "ctrl", "Todos", "action", "Create", "route", "POST /todos/")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewDeleteTodosContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Delete(rctx)
	}
	h = handleTodosOrigin(h)
	service.Mux.Handle("DELETE", "/todos/:id", ctrl.MuxHandler("delete", h, nil))
	service.LogInfo("mount", "ctrl", "Todos", "action", "Delete", "route", "DELETE /todos/:id")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewDeleteAllTodosContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.DeleteAll(rctx)
	}
	h = handleTodosOrigin(h)
	service.Mux.Handle("DELETE", "/todos/", ctrl.MuxHandler("delete all", h, nil))
	service.LogInfo("mount", "ctrl", "Todos", "action", "DeleteAll", "route", "DELETE /todos/")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewListTodosContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.List(rctx)
	}
	h = handleTodosOrigin(h)
	service.Mux.Handle("GET", "/todos/", ctrl.MuxHandler("list", h, nil))
	service.LogInfo("mount", "ctrl", "Todos", "action", "List", "route", "GET /todos/")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewShowTodosContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Show(rctx)
	}
	h = handleTodosOrigin(h)
	service.Mux.Handle("GET", "/todos/:id", ctrl.MuxHandler("show", h, nil))
	service.LogInfo("mount", "ctrl", "Todos", "action", "Show", "route", "GET /todos/:id")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewUpdateTodosContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Update(rctx)
	}
	h = handleTodosOrigin(h)
	service.Mux.Handle("PUT", "/todos/:id", ctrl.MuxHandler("update", h, nil))
	service.LogInfo("mount", "ctrl", "Todos", "action", "Update", "route", "PUT /todos/:id")
}

// handleTodosOrigin applies the CORS response headers corresponding to the origin.
func handleTodosOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "http://localhost:8080/swagger") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Expose-Headers", "X-Time")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}